Mfvargo/issue17

+++ src/dsp/power_meter.rs
@@ -25,7 +25,7 @@ impl PowerMeter {
     pub fn get_avg(&self) -> f64 {
         self.last_avg
     }
-    pub fn add_frame(&mut self, data: &[f32], gain: f32) -> () {
+    pub fn add_frame(&mut self, data: &[f32], gain: f64) -> () {
         let p = get_frame_power_in_db(data) + to_db(gain);
         self.last_peak = self.peak.get(p as f64);
         self.last_avg = self.avg.get(p as f64);
